device-management/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
redis
pika
paho-mqtt
pydantic
python-dotenv
email-validator
requests
psutil
prometheus-fastapi-instrumentator
